buildsys: stop linking kernel extensions against libgap
For a 'traditional' GAP, we didn't do that anyway; but we recently
started doing this for installed GAP versions. This was meant to
help other users of libgap, but since then I have my doubt that it
has any effect for this use case. On the other hand, it apparently
caused some issues for the Debian packaging. So let's remove it
again.