Introduce MakeIsPlistVectorRep helper
This centralizes the creation of IsPlistVectorRep instances,
reducing code duplication, and enabling future refactoring.

It also fixes bugs caused by incorrect reusing of the TypeObj of
a vector to create another vector; this introduces bugs because
the type can store things like the value of `IsZero` which must
not be carelessly copied over to a new vector with different content.