`IsVecOrMatObj` does no longer imply `IsCopyable`
The filter `IsVecOrMatObj` is not documented.
The change means that `IsVectorObj` and `IsMatrixObj` objects
are not automatically in `IsCopyable`.

The idea is that each representation of such objects
(such as `IsPlistVectorRep`) decides whether its objects
admit mutable variants or not,
by installing an implication to `IsCopyable` in the former case.

The currently available "constructing filters" of vector and matrix
objects in the GAP library behave like this, and their documentation
states that the objects are in fact copyable.