temporary fix of a vector space function
When one creates a "row space" of vector objects then the base domain
of the generators must contain the given left acting domain,
otherwise not all scalar multiples of the vectors can be created.
Thus it may be necessary to replace the given generators by vectors
whose base domain is large enough.
(Alternatively, one could signal an error.)

As far as I understand, pull request #4797 had added/changed
some code in `lib/vspcrow.gi` that makes it possible to deal with
"row vector spaces" whose elements are in `IsVectorObj` (but not lists).

A lot of functionality for that is still missing,
the current change just makes one test pass that had not failed before
because of missing consistency checks.