mat8bit: fix ZeroSameMutability
The previous behaviour of the methods installed for ZeroSameMutability was
inconsistent with the behaviour of AdditiveInverseSameMutability, this commit
makes them have the same behaviour, i.e. that if ZeroSameMutability is called
for a mutable matrix <m>, then the <m> can have mutable or immutable rows
(in any combination), but the mutability of the rows in the output was determined
by the mutability of the first row only. This commit gives the result of
ZeroSameMutability the same pattern of mutable/immutable rows as the input.