mat8bit: make TriangulizeMat compatible with OneSameMutability etc
Previously if the input was a mutable matrix, then that matrix was
changed in-place to contain its triangularized version with the same
pattern of mutable and immutable rows. This commit makes the behaviour
of TriangulizeMat compatible with that of the other unary functions for
matrices in mat8bit.gi, in that the mutability of the rows of the output
is determined by the mutabilty of the first row only in the input
(mutable = all rows in output mutable, immutable = all rows in output
immutable).