mat8bit: AdditiveInverseSM compatible with OneSM etc
Previously if the input was a mutable matrix, then the output had the
same pattern of mutable and immutable rows as the input. This commit
makes the behaviour of AdditiveInverseSameMutability compatible with
that of the other unary functions for matrices in mat8bit.gi, in that
the mutability of the rows of the output is determined by the mutabilty
of the first row only in the input (mutable = all rows in output
mutable, immutable = all rows in output immutable).