matrix/vector objects: use tag based operations not constructors
Turned `NewCompanionMatrix`, `NewIdentityMatrix`, `NewMatrix`, `NewVector`,
`NewZeroMatrix`, `NewZeroVector` from constructors to tag based operations.

The code for these operations has not been changed yet.
The next step will be the introduction of default methods,
such that some of the individual methods can be removed,
and the documentation can be changed accordingly.