WORKAROUND: Avoid problems in Fr/Kan packages
The Kan package copied library code and thus uses undocumented entries
of rewriting systems. Added workaround to avoid triggering an error
in the test.

fr package messes with the .tzrules without handling the pairs2check.

Also commented out old code that is not used anymore but could be useful to
have around for future debugging

This resolves #5478