fixed `ConsiderStructureConstants`
It may happen that `ConsiderStructureConstants` excludes all candidates.
Up to now, it was possible that one runs into an error in such a case
because the function wants to proceed and tries to access entries in an
empty list.

Benjamin Sambale found an example in which the `quick` parameter is set
to `true` and hence some candidates do not get excluded in an earlier step,
but the error could in principle happen also with the default setting
`quick:= false`.