Make GroupWithMemory behave more like Group
In particular, ensure that the wrapped generators are made
immutable, and in the case of matrices, possibly converted
to a better representation. This is achieved by calling
`GroupWithGenerators` on the given generators, and then
wrapping the generators of the resulting group.

This avoids a surprising mismatch in behavior between
ordinary groups and groups with memory in some situations.
And of course making the generators immutable prevents
further pitfalls.

Also make sure we have at least one test for GroupWithMemory
where the input is not a group.