Fix HPC-GAP crash when calling `IntFFE` repeatedly
The kernel function `INT_FF` had a bug under HPC-GAP that could lead to
it allocating more and more memory (that it did not actually need),
eventually causing GAP to come to a halt or the system running out of
memory.

This kernel function is called by e.g. `IntFFE` and other user facing
code. As a result, this code did not terminate (with the fix it runs
in a few milliseconds):

    for p in Primes do IntFFE(Z(p)); od;

For the experts: the code was directly calling `GrowPlist`, which
does not perform a check to see if the list is actually already big
enough. So each time `INT_FF` was called ended up increasing an
internal list by a certain small percentage. But doing this often
enough still leads to exponential growth of the size of that list...