Disable tests for DirectoryHome, DirectoryDesktop
These tests are not very important, but easily break -- e.g there might
be no "desktop directory". Also, guix sets HOME to an invalid directory
path *on purpose*, GAP works 100% in this setting, *except* for this
test.

In the other direction, I don't really see anything we can from this
test, so instead of trying to fiddle around with it, I decided to just
scrap it.

At the same time, let's make the code a little bit more robust in
case HOME is not even set.