Avoid passing immutable objects to Objectify
Immutable objects should never be able to turn into mutable objects.
Strictly speaking this is violated when passing an immutable list or
record to Objectify. Unfortunately we can't currently reject such inputs
because some GAP library code relies on this misfeature (and possibly
also in some packages). But we can at least reduce the occurrences for
now to prepare for a potential stricter future.

This is part of my long-term goal to minimize "magic" changes to GAP
types, which is an incredibly dangerous feature. It also makes correct
and sage bridging to other language more difficult: foreign language
wrapper objects around GAP objects always have to worry about the GAP
object magically changing. It also was always an issue for HPC-GAP.
Thus I hope that one day we can restrict this to a minimal and fully
enumerable set of possible conversions.