kernel: replace %I format by %H format in Pr
We used to only have %I which expects a C string; this was later joined
by %H which does the same but takes a GAP string. In the meantime, all
uses of %I were replaced by %H as in all cases this was necessary for GC
safety.

Since %I is for printing "identifiers" it seems sensible to keep that
format name instead of %H which was only used because H is the next
letter after I but which otherwise has no special meaning.