Replace uses of x->IsBound(x)
This is a complicated way of saying `x->true` or `ReturnTrue`. It did
seem to do something sensible when used as argument for `Filtered`, but
really, it could have been replaced by the named alternatives.

So instead of `Filtered(a,x->IsBound(x))`, either use `Compacted(a)` or
avoid the need for this operation altogether.