tst: unify some permutation constructor tests
Properly test bounds checks in the kernel for creating permutations:
one of the errors can only ever be reached in 64 bit mode; the other
however can always be reached, and thus should always be tested.

Note that the tst/test32bit and tst/test64bit directories were supposed
to be removed in 2018 (!) and files in there were in fact not loaded
since then (!!). My guess is that a file in each remained because of PRs
being merged "out of order" so that the PR which was supposed to remove
those dirs was merged only after some other PR added new files.

But we actually tested the constructors as part of the `RestrictedPerm`
tests, but it really wasn't appropriate there (the tests didn't even
reach `RestrictedPerm` as the kernel errors occur before the call).