Fix dev/audit-config-h.sh and how we use it (#5570)
Looking at our CI logs, `dev/audit-config-h.sh` actually always failed
since recently, but didn't signal this via its exit code. Specifically,
the `egrep` call failed but it runs in a subshell which hides the
error... so now we set shell options to detect errors in subshells.

Next, we need to fix the error. This was caused by dev/audit-config-h.sh
not working correctly when used outside the git root directory -- but
that's precisely what we want, as we use it in out-of-tree builds. So
fix the script, and also its invocation in dev/ci.sh, to work in this
case (e.g. by tweaking the `git grep` invocations to work in subdirs of
the git root).

Finally, we actually missed a regression were a file did not include config.h
that should. So we also fix that, as the now fixed tests of course
complain about it otherwise.