Tighten checks in Objectify to only accept plain inputs (#5554)
Also clarify the documentation -- at least for "standard" GAP; it now
misses to mention that atomic lists/records can also be used as input in
HPC-GAP, but that seems acceptable, given the state of HPC-GAP and also
given that we don't mention specifics for it most of the rest of the
regular GAP reference manual.

Background: In the distant path we allowed Objectify to be invoked on
any kind of GAP list or record, even though it was only intended for
plain lists and records. Passing in another kind of list or string could
in principle be used to "hack" any object into something different. But
that seems like a *bad* idea in general, and there is no clear use case;
however, there are cases where this can be done by accident.

Hence I tightened this check in early 2020; however, it turned out that
there was actually a call to Objectify on a plain string in HAP, so we
added a workaround for that. This patch removes that workaround.