Remove various cygwin specific files
I don't think many if any people are building GAP directly
from git on cygwin, and even if they do, I am not sure any of
these things would have been very useful...

Note that since enable_compat_mode was removed some time ago,
we didn't even build gap.bat and gapcmd.bat anymore.

The only file that seems to be in use is gaplogo.ico, but the
use is in the gap-system/gap-windows repository, so we should
just copy it there.