changed the generic `PermutationOp` method
Apply the action function to the points in the action domain,
not to images computed by the action function.
This is expected to yield a small improvement in the case of
actions on elements of a f.p. group, where the lengths of the
words tend to grow when one computes images.
(The tests from `tst/testextra/grpperm.tst` are indeed sped up
by the proposed change.)