Fix an error in evaluating the MaximalAbelianQuotient of a subgroup of an FPGroup which could result in wrong results (#5620)
* Conjugacy test for cyclic subgroups

Do not try the orbit mapping test for cyclic subgroups (or subgroups with
many orbits). Thiscould be memory intensive, and is done reasonably by
backtrack.

This resolves #5564

* FIX: Abelianized subgroup rewriting

Replace method for `MaximalAbelianQuotient` for a subgroup with an easier
version that rewrites the presentation and then abelianizes.

This fixes #5609

* ENHANCE: AGSRCharacteristicSeries refine

Only attempt refinements with PCore, Agemo, Omega, if steps are not already
elementary abelian. This can avoid expensive subcomputations.