kernel: signal file open errors in a consistent way (#5634)
... by returning -1 in SyFopen, even for the case were we run out of
internal file buffers.

Then adjust the test suite check for this situation accordingly.

This fixes intermittent failures in this test on macOS and OpenBSD: for
both of these we run out of OS file descriptors before we run out of
GAP file buffers. But that situation was previously handled differently:
`OutputTextFile` just returns `fail` for most errors opening a file,
just in the one case were we run out of internal buffers is an actual
error raised.

Besides fixing a portability problem with the test suite, I also think
this is the "right" thing overall, in terms of overall consistency.