Release scripts: fix download_with_sha256() (#5650)
This function failed to freshly download the .sha256 checksum files
since we switched the `curl` flags to include `-C` for continuing
downloads. A simple fix would be to just always delete local copies of
those files and thus force a fresh download. This goes one step further
and uses the `requests` package to fetch the checksum directly into a
string, bypassing the file system.