kernel: deal with T_DATOBJ instances without type (#5654)
Some kernel extension (EDIM, float) create T_DATOBJ instances for
use as internal buffers without ever setting a typeobj. This
mostly works as long as the buffer stays 100% local. But at least
in EDIM, one of the buffers gets retyped into a plist, which used
to be fine, except now we perform some sanity checks in
`PrecheckRetypeBag`.

Arguably the "correct" way to deal with this is to change affected
packages to use `NewKernelBuffer`. But it seems prudent to also
deal with this case directly in GAP itself. Especially since we
also still created a ton of T_DATOBJ directly in GAP and I think
we don't set a type for all of them either.