support initial search in `PAGER_BUILTIN`
When `PageSource` calls `Pager` with the data for a kernel function,
it specifies the start line by a search string.
This works well for external pagers such as `less`,
but up to now `PAGER_BUILTIN` did not honor this hint.

With the changes proposed here,
calls such as `PageSource( IS_SSORT_LIST_DEFAULT )` will work as expected
also when `PAGER_BUILTIN` is used.