fix `PreImagesRepresentative` for `IsToPcGroupHomomorphismByImages`
Added a check whether the given element is in the image of the map.
Now the method returns `fail` if not, as is specified in the
documentation of `PreImagesRepresentative`.

Note that `InversePcgs` is called (only) in the
`PreImagesRepresentative` method for a group homomorphism in
`IsToPcGroupHomomorphismByImages`.
It sets the component `rangePcgs`; the value is a pcgs for the *image*
of the homomorphism (in general *not* for the range).