Improve construction of permutation actions and by this e.g. make `FactorCosetAction` faster by several magnitudes in certain examples (#5698)
* PERFORMANCE: Permutation Group transversal

Rewrite a condition, mapping with inverse permutations, to test for a
small set of numbers rather than a large one. Doing so can give a
substantial speedup for constructing certain right transversals.

Also adjust of storage limits to reflect hardware progress over
20 years.

* ENHANCE: Orbit calculation may keep permutations

using an option, if they are obtained for free during the calculation

* ENHANCE: FactorCosetAction and RightTransversal

Speed up these operations by finding a better subgroup series, defined by
actions, and use these actions if possible.

Added example and clarify that ordering of points in `FactorCosetAction`
is not guaranteed.