fix indentation when printing double cosets
The distribution of `\<` and `\>` hints was obviously wrong.

I still do not understand why the hints were chosen this way.
Usually the idea is to mark positions in the string where line breaks
are more suitable than in other places, for example at the commas
that separate objects; but here it is the other way round?