kernel: clarify comments when calling InitGlobalBag
Several of the old comments were phrased in a way that suggested
`InitGlobalBag` would "create" something, which is not the case.
Also some referred to Gasman, even though `InitGlobalBag` also
does something when using the Julia garbage collector.

Also remove some `#ifndef HPCGAP` which were redundant as
`InitGlobalBag` already is a no-op in HPC-GAP.