kernel: enforce SyUseReadline invariant more
SyUseReadline is always 0 / false if HAVE_LIBREADLINE
is undefined.

In fact ideally it would not even be defined unless
HAVE_LIBREADLINE is defined, but we don't do that for
logistical reasons: it would affect the ABI, and would
require enforcing that `config.h` is included before
`sysopt.h`, which just is not practical at this point.