kernel: split CLI arg parsing into two stages
Most arguments are handled in the first stage. In particular all
that are relevant for the GC. The second stage is only for the
arguments which add GAP roots paths: those involve storing
strings, which can be of arbitrary size. We'd like to switch their
handling from C strings to GAP strings, but that requires the
memory manager to be initialized. With this patch, it *is*
initialized before we parse the root paths.