kernel: remove `SyOriginalArgv` & `SyOriginalArgc` (#5887)
These kept references to the `argv` array around. This is
potentially problematic for consumers of libgap API, specifically
when calling `GAP_Initialize` it is not obvious that one has to
keep the `argv` passed to it alive until after GAP completed its
startup.

While we could amend the documentation for `GAP_Initialize`
(which unfortunately currently does exist at all) to mention this,
it seems better to obviate the need for it.

This patch is a first step towards this. However, several pointers
to `argv` members remain, at least these:
- `SyCompileOutput`
- `SyCompileInput`
- `SyCompileName`
- `SyCompileMagic1`
- `SyRestoring`