Change 3 arg Assert to raise an error (#5904)
Surprisingly, the three argument version of GAP's `Assert` statement
did not raise an error, unlike the more commonly used two argument
version. That is:

    gap> Assert(0, false);
    Error, Assertion failure
    not in any function at *stdin*:1
    you may 'return;'
    brk>
    gap> Assert(0, false, "MESSAGE");
    MESSAGE

This is a really surprising an unexpected behavior. After some
discussion we agreed that both should raise an error. While this
is technically a breaking change, we expect the impact of this on
the GAP package ecosystem to be minimal, and if at all positive:
all uses of the three argument version we found all seem to be
written by someone expecting them to raise an error!

So the new behaviour now is this:

    gap> Assert(0, false, "MESSAGE");
    Error, Assertion failure: MESSAGE
    not in any function at *stdin*:1
    you may 'return;'
    brk>