Set HasBaseDomain filter for matrixobj-with-memory
... at least if the wrapped matrixobj had it set (which really should
always be the case).

This fixes a nasty long-standing issue in genss / orb / recog where
`DefaultScalarDomainOfMatrixList` returned to small a field when called on a
list of matrixobj-with-memory. This then caused `GroupWithGenerators` to
rewrite the generators over a smaller field, which then lead `genss` to
produce (compressed) vectors for use as base points in the orbit algorithm
over the smaller field; and a matching hash function. When later generators
got added that really needed the larger field, everything exploded in
confusing ways.