Load the whole GAP library before loading packages
Up to now, the "implementation part" of the GAP library was delayed
until a package got loaded that claimed to need the whole library.
The idea was to have calls to functions from some packages inside
the GPA library.
This was used for some time by the packages primgrp and transgrp,
but meanwhile the GAP library can apparently be loaded without problems
before the declaration part of these packages has been loaded.