Improve printing of nested expressions
Specifically, make it better at eliding redundant parenthesis. We
do so by changing the code so that the PrintPrecedence is reset to
0 whenever a expression is print that is not a unary or binary op,
and then restored afterwards. For "leave" expressions such as
integer literals this has no effect. But for expression that can
have further subexpressions, such as function expressions, list
expression, permutation expressions etc., this allows us to avoid
redundant parenthesis.