FIX: bug in IsomorphismGroups
The matching of characteristic subgroups matched some subgroups without
testing that they indeed can match. This could lead to an error when no
partner is found.
This issue can only come up if the groups are nonisomorphic, so it should
not be able to lead to wrong results.

This fixes #5931