Make ArithmeticElementCreator understand re-read mode
In GAP one can use `Reread` (and its variants) instead of `Read`
(and its variants) while developing code; this suppresses errors
caused by re-defining global read-only variables.

But this didn't work when using `ArithmeticElementCreator` because
that had its own check for detecting and rejecting such
redefinitions.

This patch teaches it to honor the `REREADING` global for a
smoother developer experience.