Fix a bug computing EndAlgResidue for a meataxe module
In the case of a 1-dimensional endomorphism algebra, the code
assumed that the basis of the endomorphisms would consists of
an identity matrix. But that is not always the case (we make no
attempt to "normalize" the basis elements we compute).

This fixes a bug reported by David Roe on the GAP Forum, where

    AutomorphismGroup(TransitiveGroup(45, 3878));

would sometimes return an incorrect (too small) result.