Speed up IsRegularPGroup for direct products
The direct product of two regular p-groups may or may not be
regular. Dealing with this case so far has been a worst case
scenario for the current algorithm driving `IsRegularPGroup`.

The revised algorithm in this commit uses a criterion due to
Avinoam Mann that applies as soon as a p-group has a non-cyclic
center, and which thus is perfectly suited to deal with direct
products of p-groups.

The new criterion adds some tiny overhead for smaller examples that
already worked well before, but blows it out of the water for larger
examples.

Before:

    gap> G:=DirectProduct(SmallGroup(3^5,22),SmallGroup(3^5,39));;
    gap> IsRegularPGroup(G); time;
    true
    3968

    gap> G:=DirectProduct(SmallGroup(3^5,22),SmallGroup(3^5,22));;
    gap> IsRegularPGroup(G); time;
    false
    1470

    gap> G:=DirectProduct(SmallGroup(3^5,22),SmallGroup(3^5,22),SmallGroup(3^5,24));;
    gap> IsRegularPGroup(G); time;
    false
    187669

After:

    gap> G:=DirectProduct(SmallGroup(3^5,22),SmallGroup(3^5,39));;
    gap> IsRegularPGroup(G); time;
    true
    38

    gap> G:=DirectProduct(SmallGroup(3^5,22),SmallGroup(3^5,22));;
    gap> IsRegularPGroup(G); time;
    false
    1288

    gap> G:=DirectProduct(SmallGroup(3^5,22),SmallGroup(3^5,22),SmallGroup(3^5,24));;
    gap> IsRegularPGroup(G); time;
    false
    1573

Of course one can construct even larger example which are still out of reach...

I'd like to dedicate this work to R. Keith Dennis, whom I never met in real
life but who asked a great many questions about regular p-groups on the
group-pub-forum, which eventually led to me implementing IsRegularPGroup.
I am not sure if anyone will ever use this function now that Keith is not with
us anymore, but I'd like to think he'd enjoy it if he was still around.