Add benchmarks for ExtractSubMatrix + CopySubMatrix (#5958)
These functions are part of the MatrixObj AP, but currently they
are basically always slower than their list arithmetic counter
parts involving the `m{}{}` syntax, even in situations where those
APIs were supposed to be superior.

I believe this can be reversed by providing tuned methods for these
functions (at least for the built-in plain and compressed matrices).
In order to tackler this systematically, we need robust benchmarks
to measure the effect of any changes.