Differentiate gap vs standard lib headers (#4157)
The header io.h from MinGW is hidden by gap's io.h which is causing
compilation error. Changing -I to -iquote ensures that all gap header
files are included with quotes, while the standard library header files
are included with the <>. Also, created extra/ that contains headers
to be exported for packages.