Clean up non-portable long family types
As a followup of PR #6014, we are cleaning up long family types
(including long and long double) that has different sizes in MinGW and
Linux gcc. This clean up includes all uses outside hpc, extern that
are not part of some print statements.

All of these uses of long are benign, except for the blister case and
some of the inaccurate comments. The blister changes are manually
tested in gap by adding print statement on the list lengths and
comparing huge boolean lists of size above 2^33.