CI: drop another ABI=32 test; earlier 'apt-get update'
In order to avoid issues with 'apt-get' caching, we need to invoke
'apt-get update' before attempting to install or remove any other
packages. But now we get an error with one of the ABI=32 tests:

    E: Unable to locate package libgmp-dev:i386
    E: Unable to locate package libreadline-dev:i386
    E: Unable to locate package zlib1g-dev:i386
    E: Unable to locate package expect:i386

So we drop that test. That leaves one ABI=32 test, which still
works because it builds its own GMP & zlib. This seems fair
overall given how irrelevant and untested 32bit builds are today.