CI: better support for PRs made from main GAP repository
It can be convenient for people with direct write access to the
main GAP repository to put branches there (instead of their personal
fork). For example, GitHub can automatically delete such branches
once they are merged into master.

But so far this was problematic because not all CI tests were run
on PRs created from such a branch.

This PR is changing that.