kernel & buildsys: change how internal GAP headers are included to minimize conflicts with OS headers (#6007)
The header io.h from MinGW is hidden by gap's io.h which is causing
compilation error. This commit solves the problem by removing gap
headers from -I in CPPFLAGS. Instead, we also add a directory extra/
that contains headers to be included by packages.

The other changes are all in support of this.

---------

Co-authored-by: Max Horn <max@quendi.de>