Fix the handling of arguments in `OrbitishFO`
The meaning of the arguments of the wrapper functions
constructed by `OrbitishFO` are determined by the number of arguments
together with the information whether the last argument is a function.

Thus it is not necessary to decide the meaning by checking conditions
about the relations between arguments.
Note that a recent change to these conditions broke a `Stabilizer`
computation (see the newly inserted test).