Speed up `CharacterDegrees` (#6075)
- Use `BaumClausenInfo` if the group in question is supersolvable.
- Redistribute the available code to methods.
- Fix a few typos.
- Change the delegation:
  Up to now, the one-argument version delegated to the two-argument version,
  now it is the other way round whenever possible.
  The idea is to use a stored attribute value whenever possible.
  (The results of the two-argument are still mutable.)

Computing the ordinary character degrees for the groups of order up to 100
needs more than 19 seconds with the old `CharacterDegrees`,
the proposed variant needs about 4,6 seconds.
(The speedup is bigger for samples of larger groups,
provided that sufficiently many supersolvable groups are contained.)

* adjust a test to the changed situation (the interaction between the methods has changed)

* adjust a manual example: With the proposed changes, the command `CharacterDegrees( SymmetricGroup( 4 ), 2 )`
(which occurs in a manual example) does no longer need to compute a character table, therefore the global counter is now smaller.

* add code in order to keep the behaviour that the two argument version returns mutable results